Simplify TupleData
Though extending `Seq[Data]` might be handy, we can add it when we need
it. We also determined that math does allow for an empty tuple so I
modified the constructor accordingly. As a result, the companion `apply`
with the `Seq` argument conflicts so I renamed that to `fromSeq`. Note
that the varargs in the case class definition give us the unapplySeq
behavior that we want.