Add validation flag to FdmlReader construction
In order to keep configuration out of the core library, users of the
FdmlReader must now specify whether they want to validate the fdml during
construction. The configuration option is now applied in the
FdmlDatasetResolver such that requests for a dataset of the form
`Dataset.fromName("foo")` will use the `latis.fdml.validate` config
option.