Modify SampledFunction definition
This moves the SampledFunction definition to Data.scala so we can make Data a sealed trait.
The emphasis is moved to streaming by redefining "samples" to return a Stream while "sampleSeq"
returns a Seq of Samples for MemoizedFunctions.
Operation application has changed significantly. Instead of a SampledFunction implementing
each method of the functional algebra, it simply has the option to apply an Operation. This
gives more control for operation application to the Dataset.
Function application via "apply" takes a step towards error handling with Either.