Modify Operation definitions
- UnaryOperation no longer provides a default no-op implementation to improve safety.
- StreamOperation defines an operation via a fs2.Pipe (Stream => Stream) to further
emphasize streaming. The goal is for all Operations to extend StreamOperation. This builds
on the idea of using "fold" to effectively memoize a SampledFunction into the range of a
ConstantFunction with its lone Sample still in the effectful Stream. Composition is no
longer a requirement of a StreamOperation. It is still in flux.
- Aggregation has settled on the idea of reducing a set of samples, either from a Dataset
or resulting from a GroupOperation, to a single range value which can be represented as
a ConstantFunction.
- Some methods have dropped "make" from their name to match our intended style.
- Even though the idea of allowing a MemoizedDataset to manage any Data as opposed to a
SampledFunction is working out, UnaryOperation's applyToData is back to a SampledFunction.