Add DatasetFunction to support evaluation
Instead of relying on a SampledFunction to support evaluation, a Dataset can be
converted into a DatasetFunction that provides an evaluation function from the
domain type to the range type. A DatasetFunction can also wrap a computational
function (e.g. coordinate system transform). This enables SampledFunctions and
computational functions to be used interchangeably. The Substitution operation
uses this approach.