Modify usage of TupleData
TupleData is now a first-class Data type. Code has been modified to use DomainData and RangeData in the context of Samples and SampledFuncitons. TupleData replaces their use in other places so that we can define functions as Data => Data. This has allowed the clean up of some SampledFunction evaluation (apply) methods. This also introduces safer construction methods for Data that return Either.