DataType refactor
Make DataType (a.k.a. the model) correct by construction. That means making Scalars, Tuples, and Functions with smart from* constructors that return Either as opposed to constructing them directly.

Scalar metadata is converted into type safe properties with consistency logic. The raw metadata is kept around so we can easily manage additional properties. It also serves as a way to copy a Scalar: modify it's metadata, then rebuild with fromMetadata to ensure consistency. Tuples and Functions have no metadata (for now?), just optional Identifiers.

One goal is to keep the sealed DataType trait file lean.
Construction logic is in Factory traits that are extended by the companion objects.
Behavior is defined in Algebra traits that are extended by the classes.

Another goal is exhaustive pattern matches (#304) and no compiler warnings. The only remaining warnings are for netCDF deprecations.