Allow constructing servers from HttpRoutes
Prior to this commit, servers could only be constructed using
subclasses of ServiceInterface, which contain HttpRoutes. It would be
useful for the auth use case to be able to manipulate these routes,
but as written there was no way to do this. This commit gives us a way
to manipulate routes from a service interface before using them to
build a server.